// frontend/profdiff-web-application/src/app/models/dto/benchResultsMetadata.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Metadata from the bench-results.json file, if not found then null.
 */
export interface BenchResultsMetadata {
  /**
   * Name of the benchmark suite.
   */
  benchmarkSuite?: string;
  /**
   * Name of the benchmark.
   */
  benchmarkName?: string;
  /**
   * Commit hash in GraalVM git repo used for compilations in current the benchmark.
   */
  commitHash?: string;
  /**
   * Platform (OS) on which the benchmark was run.
   */
  machinePlatform?: string;
  /**
   * Version of the GraalVM used for the benchmark.
   */
  graalVersion?: string;
  /**
   * Version of the JDK used for the benchmark.
   */
  jdkVersion?: string;
  /**
   * Time sampling periods (_NOTE_: not 1-1 relation with cpu time nor milliseconds) of the whole benchmark run, one for each metric round.
   */
  metricValues?: Array<number>;
}